Added DDT, DDR, NDDT, NDDR genomes and experiments

+++ INFOEA.Assignment1/INFOEA.Assignment1/Algorithm/Goal.cs
@@ -9,20 +9,20 @@ namespace INFOEA.Assignment1.Algorithm
     class Goal
     {
         public int MaxGenerations { get; private set; }
-        public int MinFitness { get; private set; }
+        public float MinFitness { get; private set; }
 
-        public Goal(int _max_generations, int _min_fitness)
+        public Goal(int _max_generations, float _min_fitness)
         {
             MaxGenerations = _max_generations;
             MinFitness = _min_fitness;
         }
 
-        public bool AchievedGoal(int generation, int fitness)
+        public bool AchievedGoal(int generation, float fitness)
         {
             return generation >= MaxGenerations || fitness >= MinFitness;
         }
 
-        public bool AchievedFitnessGoal(int fitness)
+        public bool AchievedFitnessGoal(float fitness)
         {
             return fitness >= MinFitness;
         }
